Made further mods for repulsion/readability
I set the max circle size to be 10% of the height of the screen and this
seems to have improved repulsion since circles aren't so large now that
they get "stuck" between more than one repulsive force. I also changed
the t2 from white to grey so that the text is readable when against the
white background when it overflows the circle. Finally, I added
reciprocal repulsion so that the system settles a bit faster and
smoother.